+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/extension/ExtensibleResourceAdapter.java
@@ -127,4 +127,9 @@ public abstract class ExtensibleResourceAdapter<T> extends ResourceAdapter<T> im
     return withTimeout(timeoutInMillis, TimeUnit.MILLISECONDS);
   }
 
+  @Override
+  public ExtensibleResource<T> unlock() {
+    return newInstance().init(resource.unlock(), client);
+  }
+
 }