+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 #### Improvements
 * Fix #5368: added support for additional ListOptions fields
+* Fix #5377: added a createOr and unlock function to provide a straight-forward replacement for createOrReplace.
 * Fix #5388: [crd-generator] Generate deterministic CRDs
 
 #### Dependency Upgrade